==> 2_RawdataSorting.R <==
# This script allows to select a subset of the data at each false positive tolerance using
# the corresponding confidence score calculated previoulsy (script 1_ErrorRiskModelling.R) 
# for each species.
#############################################################################################
library(boot) # version 1.3-20
library(lme4) # version 1.1-18-1
library(data.table) # 1.11.4
#############################################################################################
# By Kévin Barré, Julie Pauwels and Yves Bas
#############################################################################################

rm(list=ls())

library(readr) # vesrion 1.1.1
library(data.table) # version 1.11.4

setwd("./") # Your work directory

# Reading the file saved in the script 1.
ErrorRiskThresholds <- read.csv("./ErrorRiskThresholds.csv")
# Adding a column for the calcultation on raw data
ER0 = seq(0,0,length.out=nrow(ErrorRiskThresholds))
ErrorRiskThresholds = cbind(ErrorRiskThresholds, ER0)

# Reading the raw file generated by the automatic identification software, containing one bat pass
# per row assigned to a species with a confidence score of the identification
#It must contain at least those columns:
  # - the site ID ("ID")
  # - the species identified by the software ("species")
  # - the confidence score of the identification ("confidence_index")
datatot <- fread("./data_total.csv")
datatot$species = as.character(datatot$species)

# Removing of the noise and species absent from the manual checks dataset
datatot = datatot[-which(datatot$species %in% c("noise", "Vesmur", "Tadten", "Pippyg", "Minsch", "Hypsav")), ]

# Grouping of myotis and plecotus species into two groups (Myotis spp. and Plecotus spp.)
# because the manual identification of these species is too uncertain (but you can adapt it to your study aims)
# We keep the original "species" column to keep Myotis nattereri, easier to identify
datatot = cbind(datatot, simp_species = datatot$species)
datatot$simp_species = as.character(datatot$simp_species)
pb <- txtProgressBar(min = 0, max = nrow(datatot), style = 3)
for (i in 1:nrow(datatot)) {
  setTxtProgressBar(pb, i)
  if (datatot$simp_species[i] %in% c("Myoalc", "Myocap","Myodau", "MyoGT", "Myomys", "Myonat", "Myoema")) datatot$simp_species[i] = "Myosp"
  if (datatot$simp_species[i] %in% c("Pleaur", "Pleaus")) datatot$simp_species[i] = "Plesp"
}
close(pb)
datatot$species = as.factor(datatot$species)
datatot$simp_species = as.factor(datatot$simp_species)
write.csv(datatot, "./data_total_speciesgroup.csv", row.names=F)

# Vector creation of different thresholds of the false positive tolerance 
# to associate it to each species
thresholds <- c("ER0","ER50", "ER60", "ER70", "ER80", "ER90")

## Bat activity dataframe creation filtering from datatot bat passes per site per night
## according to each threshold of false positive tolerance (i.e. needed confidence scores) 
## for each species

# dataframe size
ncolumn <- length(unique(datatot$simp_species))*length(thresholds) + length(thresholds)
crosstable <- matrix(nrow = length(unique(datatot$ID)), 
                     ncol = ncolumn) 
# Imputation of site ID
rownames(crosstable) <- unique(datatot$ID) 

# Column names creation (SpeciesThreshold)
col <- c()
for (i in unique(datatot$simp_species)) {
  for (s in thresholds) {
    col = c(col, paste(i, s, sep="_"))
  }
}
for (s in thresholds) {
  col = c(col, paste("Myonat", s, sep="_"))
}
colnames(crosstable) <- col

# Imputation of the number of bat passes according to the species for each threshold
for (i in unique(datatot$simp_species)) 
     {
  print(i)
  start = Sys.time()
  for (s in thresholds) 
    {
    IC = ErrorRiskThresholds[which(ErrorRiskThresholds$Species == i), s]
    for (j in unique(datatot$ID)) 
         {
      data = datatot[which((datatot$simp_species == i) & (datatot$ID == j) & (datatot$confidence_index > IC)),]
      crosstable[j, paste(i, s, sep='_')] = nrow(data)
    }
  }
  print(Sys.time() - start)
}

# Adding Myotis nattereri from the "species" column ("simp_species" only contains the "Myotis spp" level)
for (s in thresholds){
  IC = ErrorRiskThresholds[which(ErrorRiskThresholds$Species == "Myonat"), s]
  for (j in unique(datatot$ID)) 
  {
    data = datatot[which((datatot$species == "Myonat") & (datatot$ID == j) & (datatot$confidence_index > IC)),]
    crosstable[j, paste("Myonat", s, sep='_')] = nrow(data)
  }
}

View(crosstable)
# Saving
write.csv(crosstable, "./CrossTableFiltered.csv", row.names=T)
