==> 4_ErrorRate.R <==
# This script allows to compute real false positive and generated false negative rates for each species  and 
# for each false positive tolerance using the error rate modelling performed in the script 
# 1_ErrorRiskModelling.R.
#############################################################################################
# By Kévin Barré, Julie Pauwels and Yves Bas
#############################################################################################

rm(list = ls())

library(data.table) # version 1.11.4
library(boot) # version 1.3-20

setwd("./") # Your work directory

# Output of the automatic identification software
datatot <- fread("./data_total_speciesgroup.csv")

# Reading the file saved in the script 1 containing confidence indexes corresponding to each 
# false positive tolerance (50%, 40%, 30%, 20%, 10%) 
ErrorRiskThresholds <- read.csv("./ErrorRiskThresholds.csv")

# Merging both files to get thresholds for each species in datatot
data = as.data.frame(merge(datatot, ErrorRiskThresholds[,c(1:8)],by.x="simp_species",by.y="Species"))
# Adding a column for the calcultation on raw data
ERraw = seq(0,0,length.out=nrow(datatot))
data = cbind(data, ERraw)

tabErr = c()
tabFN = c()
tabTP = c()
Thresholds = c("ERraw","ER50", "ER40", "ER30", "ER20", "ER10")

# Calculation of real false positive rates for each species at each thresholds (i.e. each false
# positive tolerance)
for (i in unique(as.factor(data$simp_species))){
  # Subset on the species
  dataSp=subset(data, data$simp_species==i)
  ErrorRateTotal = c()
  FNRateTotal = c()
  FNrate = c()
  for (j in Thresholds) {
    # Subset on the threshold
    dataErr=subset(dataSp,dataSp$confidence_index>dataSp[,j])
    dataTp=subset(dataSp,dataSp$confidence_index<dataSp[,j])
    ErrorRate = 0
    TPrate = 0
    Err = 0
    TP = 0
    # Calculation of error and success probability corresponding to each confience score from the equation modelling
    # the error risk in the sript 1 on each rows (Err column) of the subset data
    Err = 1-inv.logit(dataErr$Int+dataErr$Estimate*dataErr$confidence_index) # False positive probabiity
    TP = inv.logit(dataTp$Int+dataTp$Estimate*dataTp$confidence_index) # True positive probability
    # Averaging of all real rates to estimate false positive rate
    ErrorRate = sum(Err)/nrow(dataErr)
    # Averaging of all real rates to estimate true positive rate
    FNrate = sum(TP)/nrow(dataSp)
    # False negative rate generated by discarding data under given thresholds
    if (j == "ERraw"){FNrate = 0 # False negative rate is null in whole dataset (no threshold "ER0")
    } else {FNrate}
    ErrorRateTotal = c(ErrorRateTotal, round(ErrorRate,4))
    FNRateTotal = c(FNRateTotal, round(FNrate, 4))
    rm(ErrorRate, Err, TP,FNrate)
  }
  tabErr = rbind(tabErr, ErrorRateTotal)
  tabFN = rbind(tabFN, FNRateTotal)
}

# Adding Myotis nattereri from the "specie" column ("simp_species" only contains the "Myotis spp" level)
dataSp=subset(data, data$species=="Myonat")
ErrorRateTotal = c()
FNRateTotal = c()
FNrate = c()

for (j in Thresholds) {
  dataErr=subset(dataSp,dataSp$confidence_index>dataSp[,j])
  dataTp=subset(dataSp,dataSp$confidence_index<dataSp[,j])
  ErrorRate = 0
  TPrate = 0
  Err = 0
  TP = 0
  Err = 1-inv.logit(dataErr$Int+dataErr$Estimate*dataErr$confidence_index)
  TP = inv.logit(dataTp$Int+dataTp$Estimate*dataTp$confidence_index)
  ErrorRate = sum(Err)/nrow(dataErr)
  FNrate = sum(TP)/nrow(dataSp)
  if (j == "ERraw"){FNrate = 0
  } else {FNrate}
  ErrorRateTotal = c(ErrorRateTotal, round(ErrorRate,4))
  FNRateTotal = c(FNRateTotal, round(FNrate, 4))
  rm(ErrorRate, Err)
}
tabErr = rbind(tabErr, ErrorRateTotal)
tabFN = rbind(tabFN, FNRateTotal)

colnames(tabErr) = Thresholds
colnames(tabFN) = Thresholds
row.names(tabErr) = c(unique(as.character(data$simp_species)), "Myonat")
row.names(tabFN) = c(unique(as.character(data$simp_species)), "Myonat")
# Saving
write.csv(tabErr, "./FalsePositiveRate_datatotal.csv", row.names=T)
write.csv(tabFN, "./FalseNegativeRate_datatotal.csv", row.names=T)
